Action delete API rework - support for action file deletion

+++ st2api/st2api/controllers/v1/actions.py
@@ -206,7 +206,7 @@ class ActionsController(resource.ContentPackResourceController):
 
         return action_api
 
-    def delete(self, ref_or_id, requester_user):
+    def delete(self, files_remove_request, ref_or_id, requester_user):
         """
         Delete an action.
 
@@ -240,6 +240,7 @@ class ActionsController(resource.ContentPackResourceController):
         pack_name = action_db["pack"]
         entry_point = action_db["entry_point"]
         metadata_file = action_db["metadata_file"]
+        remove_files = files_remove_request.remove_files
 
         try:
             Action.delete(action_db)
@@ -252,20 +253,30 @@ class ActionsController(resource.ContentPackResourceController):
             )
             abort(http_client.INTERNAL_SERVER_ERROR, six.text_type(e))
             return
-        try:
-            delete_action_files_from_pack(
-                pack_name=pack_name,
-                entry_point=entry_point,
-                metadata_file=metadata_file,
-            )
-        except Exception as e:
-            LOG.error(
-                "Exception encountered during deleting resource files from disk."
-                "Exception was %s",
-                e,
-            )
-            abort(http_client.INTERNAL_SERVER_ERROR, six.text_type(e))
-            return
+
+        if remove_files:
+            try:
+                delete_action_files_from_pack(
+                    pack_name=pack_name,
+                    entry_point=entry_point,
+                    metadata_file=metadata_file,
+                )
+            except PermissionError as e:
+                LOG.error("No permission to delete resource files from disk.")
+                action_db.id = None
+                Action.add_or_update(action_db)
+                abort(http_client.FORBIDDEN, six.text_type(e))
+                return
+            except Exception as e:
+                LOG.error(
+                    "Exception encountered during deleting resource files from disk. "
+                    "Exception was %s",
+                    e,
+                )
+                action_db.id = None
+                Action.add_or_update(action_db)
+                abort(http_client.INTERNAL_SERVER_ERROR, six.text_type(e))
+                return
 
         extra = {"action_db": action_db}
         LOG.audit("Action deleted. Action.id=%s" % (action_db.id), extra=extra)
